+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Expose method `EnsureCRDs` to register CRDs in the k8s API.
+- A custom `Scheme` can be passed to configure the controller-runtime client.
+- Add getter method that returns the controller-runtime client.
 
 ## [0.4.1] - 2020-10-30